requirements.txt
django
djangorestframework
simplejson
requests

